Final Temple Build,Script Bugs Fixed, Polished Mechanics
This is the Temple Build. I worked through bugs and Controls and mechanics so they were as smooth as I had time to make them. There are still things I need to add like layers for each of the "color keys" so they only open one door but overall it should all work.